job-app-tracker/requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
openai>=1.0
plotly
kaleido
numpy
pandas
tenacity>=8.0
pytest>=7.0